Taking the browser history indexer outside so the user can do it separately
